+++ src/ftw_cli/cfg.py
@@ -4,7 +4,7 @@ COLLECTION_ID = "sentinel-2-l2a"
 BANDS_OF_INTEREST = ["B04", "B03", "B02", "B08"]
 
 # Supported file formats for the inference polygon command
-SUPPORTED_POLY_FORMATS_TXT = "Available file extensions: .parquet (GeoParquet, fiboa-compliant), .fgb (FlatGeoBuf), .gpkg (GeoPackage), .geojson and .json (GeoJSON)"
+SUPPORTED_POLY_FORMATS_TXT = "Available file extensions: .parquet (GeoParquet, fiboa-compliant), .fgb (FlatGeoBuf), .gpkg (GeoPackage), .geojson / .json / .ndjson (GeoJSON)"
 
 # List of all available countries
 ALL_COUNTRIES = [